+++ passenger-service/src/main/java/io/voitovich/yura/passengerservice/exceptionhandler/ProfileExceptionHandler.java
@@ -3,6 +3,8 @@ package io.voitovich.yura.passengerservice.exceptionhandler;
 import io.voitovich.yura.passengerservice.exception.NoSuchRecordException;
 import io.voitovich.yura.passengerservice.exception.NotUniquePhoneException;
 import io.voitovich.yura.passengerservice.exception.NotValidUUIDException;
+import io.voitovich.yura.passengerservice.exceptionhandler.model.ExceptionInfo;
+import io.voitovich.yura.passengerservice.exceptionhandler.model.ValidationExceptionInfo;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
@@ -12,7 +14,6 @@ import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
