+++ passenger-service/src/main/java/io/voitovich/yura/passengerservice/exceptionhandler/model/ExceptionInfo.java
@@ -1,14 +1,18 @@
-package io.voitovich.yura.passengerservice.exceptionhandler;
+package io.voitovich.yura.passengerservice.exceptionhandler.model;
 
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NonNull;
 import org.springframework.http.HttpStatus;
 
 
-@Getter
+@Builder
 public class ExceptionInfo {
+
     private final int code;
+    @NonNull
     private final String message;
+    @NonNull
     private final HttpStatus status;
 
     public ExceptionInfo(String message, HttpStatus status) {
